Added pod documentation to XMLTV.pm.  As with tv_grep, part of it is automatically generated from the list of 'handlers' for subelements of <programme> and <channel>.  This necessitated moving XMLTV.pm to XMLTV.pm.in and writing a short script XMLTV.pm.PL to generate the end file.
Also I needed to change the way the list of handlers is stored - now they're
looked up by a string rather than being subroutine references.  This is so
we can also use the strings in the pod documentation :-).  The function
list_programme_keys() has been changed to just return the list of handlers
(more or less), which has meant changing tv_grep.  Found and fixed a bug
in tv_grep in the meantime, and made a small change to XMLTV.pm's handling
of subtitles.  Apart from those fixes and any new bugs introduced, the
behaviour should be the same - just we now have some rather pedantic
half-auto-generated documentation as well.