This is it, the new version of the UK grabber.  This fetches XML documents from Ananova's server, munges them a bit and outputs them.  It hasn't had quite as much testing as I would have liked, but because the old listings feed stopped I have to commit this now.  Also the code is not as clean as it could be, but ditto.
Another problem is slowness and bloatedness; it parses the XML received and
this means a lot of memory usage with the current XMLTV.pm.  Still, on a
fast machine the main bottleneck will be speed of download from the server
(which is very slow), so I don't feel too guilty.