After spending several hours last night looking worriedly at the modem lights and thinking 'hmm, I didn't know tv_grab_uk was this slow, even on a 28.8 connection and P200', I have improved the status messages so there's more description of what is happening.  Most important among these improvements is making the progress bar detection actually work - previously it wouldn't detect that Term::ProgressBar was installed and would remain silent.  Also, downloading now happens first of all with parsing later, so you can go offline a bit sooner.
The real reason it took so long is that I had pressed Return to get
the default 'yes' answer to 'download radio listings'.  As I announced
on xmltv-users soon after 0.5 was released, this option is stupid
because the new Ananova site has hundreds of radio channels.  Now you
must download them individually.