When prompting for channels to install, show the user the display name not the XMLTV id.  Better warning if a channel object doesn't have a main display name (shouldn't happen, but it might).
init_channels(): set the type of the newly-created channel to
'tentative'.  Then if the type is not updated later, we know that this
channel appears only in the config file and warn about its removal.

get_channel_pkgs(): the hash mapping package name to channels now
gives a list of channel objects rather than just the XMLTV ids.  We
also take the opportunity to check that the XMLTV id in the package
file is known (this means that any channel mentioned in a package must
also be listed in the channel_ids file so it is known about this
early, but that shouldn't be a problem).

read_channels_page(): once we've read the channesl from the site we
know that any channel objects we haven't updated are 'phantom'
channels mentioned in the config file but not found on the site.  This
function now removes them from the global @ch_all list and from the
channel packages set earlier (and complains about it).