As mentioned on xmltv-devel, changing the names of all the grabber configuration files to end in '.conf'.  This is so that on Windows, when run from the current directory, they do not try to overwrite themselves.  Thanks to Davide Chiarini for pointing this out.  Also it's probably more obvious what the files are for with '.conf' stuck on the end of the filename.
The migration should be automatic, you just get a message saying that
the old file is being renamed to end in .conf.  Of course _this_ will
go wrong for Windows users running from the current directory, but
it's only a short-term measure.

Hmm, perhaps some of this config file reading/writing could be
factored out.  It's a lot of almost identical code.