add_stop(): Rewrote using a simpler algorithm: go through all the following programmes and pick the start time if it is later than ours, or the stop time if it exists.  If neither condition is true move on to the next following programme.
I don't know why I made it so complex to start with, I think it was
because I thought it sensible to start at the end of the sorted list
and work backwards.  And it would be sensible, if you only looked at
start times.  But we don't any more.

This should fix the pesky bug where tv_sort --by-channels doesn't work
because add_stop() would break the sortedness of its input programmes
list.